+++ apps/ymir-sdl3/src/app/ui/windows/debug/memory_viewer_window.cpp
@@ -58,6 +58,9 @@ void MemoryViewerWindow::DrawContents() {
     ImGui::TextUnformatted("Region");
 
     ImGui::Checkbox("Enable side-effects", &m_memViewState->enableSideEffects);
+    if (ImGui::SmallButton("Dump memory region")) {
+        m_context.EnqueueEvent(events::emu::DumpMemRegion(*m_memViewState));
+    };
     if (currRegion.paramsFn) {
         currRegion.paramsFn(m_memViewState.get());
     }